Move address mapping test out to ethereum
